Share Custom data fixes from Membership form to renewal
This
addresses php8.x issues as well as consolidating code.
